+++ backend/src/main/kotlin/bo/edu/umsa/backend/controller/ReplacedPartController.kt
@@ -25,7 +25,7 @@ class ReplacedPartController @Autowired constructor(
     fun getReplacedPartById(@PathVariable replacedPartId: Long): ResponseEntity<ResponseDto<ReplacedPartDto>> {
         logger.info("Starting the API call to get the replaced part by id")
         logger.info("GET /api/v1/replaced-parts/$replacedPartId")
-        AuthUtil.verifyAuthTokenHasRole("VER TAREAS")
+        AuthUtil.verifyAuthTokenHasPermission("VER TAREAS")
         val replacedPart: ReplacedPartDto = replacedPartService.getReplacedPartById(replacedPartId)
         logger.info("Success: Replaced part retrieved")
         return ResponseEntity(ResponseDto(true, "Reemplazo recuperado", replacedPart), HttpStatus.OK)
@@ -35,7 +35,7 @@ class ReplacedPartController @Autowired constructor(
     fun createReplacedPart(@RequestBody newReplacedPartDto: NewReplacedPartDto): ResponseEntity<ResponseDto<Nothing>> {
         logger.info("Starting the API call to create the replaced part")
         logger.info("POST /api/v1/replaced-parts")
-        AuthUtil.verifyAuthTokenHasRole("VER TAREAS")
+        AuthUtil.verifyAuthTokenHasPermission("VER TAREAS")
         replacedPartService.createReplacedPart(newReplacedPartDto)
         logger.info("Success: Replaced part created")
         return ResponseEntity(ResponseDto(true, "El reemplazo se ha creado", null), HttpStatus.CREATED)
@@ -48,7 +48,7 @@ class ReplacedPartController @Autowired constructor(
     ): ResponseEntity<ResponseDto<Nothing>> {
         logger.info("Starting the API call to update the replaced part")
         logger.info("PUT /api/v1/replaced-parts/$replacedPartId")
-        AuthUtil.verifyAuthTokenHasRole("VER TAREAS")
+        AuthUtil.verifyAuthTokenHasPermission("VER TAREAS")
         replacedPartService.updateReplacedPart(replacedPartId, newReplacedPartDto)
         logger.info("Success: Replaced part updated")
         return ResponseEntity(ResponseDto(true, "El reemplazo se ha actualizado", null), HttpStatus.OK)
@@ -58,7 +58,7 @@ class ReplacedPartController @Autowired constructor(
     fun deleteReplacedPart(@PathVariable replacedPartId: Long): ResponseEntity<ResponseDto<Nothing>> {
         logger.info("Starting the API call to delete the replaced part")
         logger.info("DELETE /api/v1/replaced-parts/$replacedPartId")
-        AuthUtil.verifyAuthTokenHasRole("VER TAREAS")
+        AuthUtil.verifyAuthTokenHasPermission("VER TAREAS")
         replacedPartService.deleteReplacedPart(replacedPartId)
         logger.info("Success: Replaced part deleted")
         return ResponseEntity(ResponseDto(true, "El reemplazo se ha eliminado", null), HttpStatus.OK)
