+++ backend/src/main/kotlin/bo/edu/umsa/backend/controller/UserController.kt
@@ -22,7 +22,7 @@ class UserController @Autowired constructor(private val userService: UserService
     fun getAllUsers(): ResponseEntity<ResponseDto<List<UserPartialDto>>> {
         logger.info("Starting the API call to get all users")
         logger.info("GET /api/v1/users/all")
-        AuthUtil.verifyAuthTokenHasRoles(listOf("GESTIONAR ROLES Y PERMISOS", "CREAR PROYECTOS", "EDITAR PROYECTOS").toTypedArray())
+        AuthUtil.verifyAuthTokenHasPermissions(listOf("GESTIONAR ROLES Y PERMISOS", "CREAR PROYECTOS", "EDITAR PROYECTOS").toTypedArray())
         val users: List<UserPartialDto> = userService.getAllUsers()
         logger.info("Success: All users retrieved")
         return ResponseEntity(ResponseDto(true, "Usuarios recuperados", users), HttpStatus.OK)
@@ -38,7 +38,7 @@ class UserController @Autowired constructor(private val userService: UserService
     ): ResponseEntity<ResponseDto<Page<UserPartialDto>>> {
         logger.info("Starting the API call to get the users")
         logger.info("GET /api/v1/users")
-        AuthUtil.verifyAuthTokenHasRole("VER USUARIOS")
+        AuthUtil.verifyAuthTokenHasPermission("VER USUARIOS")
         val users: Page<UserPartialDto> = userService.getUsers(sortBy, sortType, page, size, keyword)
         logger.info("Success: Users retrieved")
         return ResponseEntity(ResponseDto(true, "Usuarios recuperados", users), HttpStatus.OK)
@@ -48,7 +48,7 @@ class UserController @Autowired constructor(private val userService: UserService
     fun getUserById(@PathVariable userId: Long): ResponseEntity<ResponseDto<UserDto>> {
         logger.info("Starting the API call to get the user")
         logger.info("GET /api/v1/users/{userId}")
-        AuthUtil.verifyAuthTokenHasRole("VER USUARIOS")
+        AuthUtil.verifyAuthTokenHasPermission("VER USUARIOS")
         val user: UserDto = userService.getUserById(userId)
         logger.info("Success: User retrieved")
         return ResponseEntity(ResponseDto(true, "Usuario recuperado", user), HttpStatus.OK)
@@ -59,7 +59,7 @@ class UserController @Autowired constructor(private val userService: UserService
     fun createUser(@RequestBody newUserDto: NewUserDto): ResponseEntity<ResponseDto<Nothing>> {
         logger.info("Starting the API call to create the user")
         logger.info("POST /api/v1/users")
-        AuthUtil.verifyAuthTokenHasRole("CREAR USUARIOS")
+        AuthUtil.verifyAuthTokenHasPermission("CREAR USUARIOS")
         userService.createUser(newUserDto)
         logger.info("Success: User created")
         return ResponseEntity(ResponseDto(true, "El usuario se ha creado", null), HttpStatus.CREATED)
@@ -72,7 +72,7 @@ class UserController @Autowired constructor(private val userService: UserService
     ): ResponseEntity<ResponseDto<Nothing>> {
         logger.info("Starting the API call to update the user")
         logger.info("PUT /api/v1/users/{userId}")
-        AuthUtil.verifyAuthTokenHasRole("EDITAR USUARIOS")
+        AuthUtil.verifyAuthTokenHasPermission("EDITAR USUARIOS")
         userService.updateUser(userId, profileDto)
         logger.info("Success: User updated")
         return ResponseEntity(ResponseDto(true, "El usuario se ha actualizado", null), HttpStatus.OK)
@@ -82,7 +82,7 @@ class UserController @Autowired constructor(private val userService: UserService
     fun deleteUser(@PathVariable userId: Long): ResponseEntity<ResponseDto<Nothing>> {
         logger.info("Starting the API call to delete the user")
         logger.info("DELETE /api/v1/users/{userId}")
-        AuthUtil.verifyAuthTokenHasRole("EDITAR USUARIOS")
+        AuthUtil.verifyAuthTokenHasPermission("EDITAR USUARIOS")
         userService.deleteUser(userId)
         logger.info("Success: User deleted")
         return ResponseEntity(ResponseDto(true, "El usuario se ha eliminado", null), HttpStatus.OK)
@@ -92,7 +92,7 @@ class UserController @Autowired constructor(private val userService: UserService
     fun getProfilePicture(@PathVariable userId: Long): ResponseEntity<ByteArray> {
         logger.info("Starting the API call to get the profile picture")
         logger.info("GET /api/v1/users/{userId}/profile-picture")
-        AuthUtil.verifyAuthTokenHasRole("VER USUARIOS")
+        AuthUtil.verifyAuthTokenHasPermission("VER USUARIOS")
         val profilePicture: FileDto = userService.getProfilePicture(userId)
         val headers = HttpHeaders()
         headers.contentType = MediaType.parseMediaType(profilePicture.contentType)
@@ -120,34 +120,34 @@ class UserController @Autowired constructor(private val userService: UserService
     ): ResponseEntity<ResponseDto<Nothing>> {
         logger.info("Starting the API call to update the profile picture")
         logger.info("PUT /api/v1/users/{userId}/profile-picture")
-        AuthUtil.verifyAuthTokenHasRole("EDITAR USUARIOS")
+        AuthUtil.verifyAuthTokenHasPermission("EDITAR USUARIOS")
         userService.uploadProfilePicture(userId, file)
         logger.info("Success: Profile picture updated")
         return ResponseEntity(ResponseDto(true, "Foto de perfil actualizada", null), HttpStatus.OK)
     }
 
 
-    @GetMapping("/{userId}/groups")
-    fun getGroupsByUserId(@PathVariable userId: Long): ResponseEntity<ResponseDto<List<GroupDto>>> {
-        logger.info("Starting the API call to get the groups by user id")
-        logger.info("GET /api/v1/users/{userId}/groups")
-        AuthUtil.verifyAuthTokenHasRole("GESTIONAR ROLES Y PERMISOS")
-        val groups: List<GroupDto> = userService.getGroupsByUserId(userId)
-        logger.info("Success: Groups retrieved")
-        return ResponseEntity(ResponseDto(true, "Grupos recuperados", groups), HttpStatus.OK)
+    @GetMapping("/{userId}/roles")
+    fun getRolesByUserId(@PathVariable userId: Long): ResponseEntity<ResponseDto<List<RoleDto>>> {
+        logger.info("Starting the API call to get the roles by user id")
+        logger.info("GET /api/v1/users/{userId}/roles")
+        AuthUtil.verifyAuthTokenHasPermission("GESTIONAR ROLES Y PERMISOS")
+        val roles: List<RoleDto> = userService.getRolesByUserId(userId)
+        logger.info("Success: Roles retrieved")
+        return ResponseEntity(ResponseDto(true, "Roles recuperados", roles), HttpStatus.OK)
     }
 
-    @PostMapping("/{userId}/groups")
-    fun addGroupsToUser(
+    @PostMapping("/{userId}/roles")
+    fun addRolesToUser(
         @PathVariable userId: Long,
-        @RequestBody groupIds: Map<String, List<Long>>
+        @RequestBody roleIds: Map<String, List<Long>>
     ): ResponseEntity<ResponseDto<Nothing>> {
-        logger.info("Starting the API call to add groups to user")
-        logger.info("POST /api/v1/users/{userId}/groups")
-        AuthUtil.verifyAuthTokenHasRole("GESTIONAR ROLES Y PERMISOS")
-        userService.addGroupsToUser(userId, groupIds["groupIds"]!!)
-        logger.info("Success: Groups added to user")
-        return ResponseEntity(ResponseDto(true, "Grupos agregados al usuario", null), HttpStatus.CREATED)
+        logger.info("Starting the API call to add roles to user")
+        logger.info("POST /api/v1/users/{userId}/roles")
+        AuthUtil.verifyAuthTokenHasPermission("GESTIONAR ROLES Y PERMISOS")
+        userService.addRolesToUser(userId, roleIds["roleIds"]!!)
+        logger.info("Success: Roles added to user")
+        return ResponseEntity(ResponseDto(true, "Roles agregados al usuario", null), HttpStatus.CREATED)
     }
 
 }
